requirements.txt
absl-py==2.0.0
accelerate==0.26.1
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.2.0
aiosqlite==0.18.0
annotated-types==0.6.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
bleach==4.1.0
blis==0.7.11
Bottleneck==1.3.5
brotlipy==0.7.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.2
confection==0.1.4
contextlib2==21.6.0
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.8
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
docstring-parser==0.15
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
gmpy2==2.1.2
huggingface-hub==0.21.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.25.0
ipython==8.15.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-server==1.23.4
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
kaggle==1.6.12
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.14
munkres==1.1.4
murmurhash==1.0.10
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
peft==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
platformdirs==3.10.0
ply==3.11
preshed==3.0.9
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0
pyzmq==23.2.0
regex==2023.10.3
requests==2.31.0
responses==0.13.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
sacremoses==0.0.43
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.11.3
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
seqeval==1.2.2
setproctitle==1.3.3
setuptools==68.0.0
shortuuid==1.0.11
shtab==1.7.0
sip==6.6.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.2.0
sympy==1.11.1
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==2.2.0
timedelta==2020.12.3
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
transformers==4.38.1
triton==2.1.0
trl==0.7.11
typer==0.9.0
typing_extensions==4.7.1
tyro==0.7.3
tzdata==2023.3
urllib3==1.26.16
wandb==0.13.4
wasabi==1.1.2
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
xxhash==2.0.2
y-py==0.5.9
yarl==1.8.1
ypy-websocket==0.8.2
zipp==3.11.0
